refactor: improve search_events OpenTelemetry semantics and module structure
## Key Improvements

### 1. Enhanced OpenTelemetry Semantic Conventions Support
- Added comprehensive OpenTelemetry semantic conventions (2025 stable)
- Proper disambiguation: "agent calls" → gen_ai.* attributes (not mcp.tool.*)
- Dynamic semantic lookup based on query content
- Support for core namespaces: gen_ai.*, db.*, http.*, rpc.*, messaging.*, k8s.*

### 2. Refactored Module Structure
- Converted monolithic search-events.ts (1385 lines) into organized module
- Clear separation of concerns across 8 focused files
- Added comprehensive documentation in CLAUDE.md

### 3. Module Architecture
```
search-events/
├── index.ts                      # Export handler (public API)
├── handler.ts                    # Main tool definition and orchestration
├── agent.ts                      # AI query translation engine
├── config.ts                     # Dataset configurations and field definitions
├── formatters.ts                 # Result formatting for different datasets
├── utils.ts                      # Helper functions and utilities
├── tools/                        # AI agent tools (for embedded agent)
│   └── otel-semantics-lookup.ts  # OpenTelemetry semantics lookup tool
└── CLAUDE.md                     # Architecture documentation
```

### 4. Enhanced AI Agent Capabilities
- Dynamic prompt enhancement based on query content
- Semantic attribute lookup for common patterns
- Better error handling and validation
- Proper field type validation for aggregate functions

### 5. Problem Solved
Original issue: "top 10 agent call spans by usage in mcp-server" incorrectly used
`has:mcp_server.tool` instead of proper OpenTelemetry GenAI conventions.

Now correctly maps to `has:gen_ai.system` for AI/LLM/agent calls.

## Testing
- All existing tests pass (13 passed, 1 skipped)
- TypeScript compilation successful
- No breaking changes to public API

## Impact
- Better semantic understanding for agent-related queries
- More maintainable and extensible codebase
- Clear documentation for future development
- Proper OpenTelemetry standards compliance
