+++ packages/mcp-server/src/tools/search-events/index.ts
@@ -1,0 +1,2 @@
+// Export the search-events handler
+export { default } from "./handler";
